requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
scikit-surprise==1.1.4
streamlit==1.37.0
joblib==1.4.2
